requirements.txt
# Core ML — installed separately in setup.sh with CUDA 12.4 index
# torch, torchvision, torchaudio → pip install ... --index-url https://download.pytorch.org/whl/cu124
transformers>=4.57.0
accelerate>=0.30.0
tokenizers>=0.19.0

# Compression
bitsandbytes>=0.43.0
auto-gptq>=0.7.1
autoawq>=0.2.5
llmcompressor>=0.4.0

# Evaluation
lm-eval>=0.4.3
datasets>=2.19.0
evaluate>=0.4.1
nltk>=3.8.1

# Vision
Pillow>=10.3.0
opencv-python-headless>=4.9.0

# Profiling & monitoring
pynvml>=11.5.0
psutil>=5.9.8
py3nvml>=0.2.7

# Data & visualization
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
scipy>=1.13.0
plotly>=5.21.0

# Notebook
jupyter>=1.0.0
ipywidgets>=8.1.0

# ONNX — onnxruntime-gpu 1.18+ supports CUDA 12.x
onnx>=1.16.0
onnxruntime-gpu>=1.18.0
optimum[onnxruntime-gpu]>=1.20.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.1
jsonlines>=4.0.0
rich>=13.7.0
typer>=0.12.0
loguru>=0.7.2
einops>=0.8.0
timm>=1.0.3
sentencepiece>=0.2.0
